+++ internal/utils/errors.go
@@ -2,7 +2,6 @@ package utils
 
 import (
 	"fmt"
-	"strings"
 )
 
 type AnalysisError struct {
@@ -66,52 +65,4 @@ func NewItemCleaningError(module, itemPath, message string, cause error) *Cleani
 		Cause:    cause,
 		ItemPath: itemPath,
 	}
-}
-
-type ErrorHandler struct {
-	errors []error
-}
-
-func NewErrorHandler() *ErrorHandler {
-	return &ErrorHandler{errors: make([]error, 0)}
-}
-
-func (h *ErrorHandler) Add(err error) {
-	if err != nil {
-		h.errors = append(h.errors, err)
-	}
-}
-
-func (h *ErrorHandler) AddIf(err error, condition bool) {
-	if err != nil && condition {
-		h.errors = append(h.errors, err)
-	}
-}
-
-func (h *ErrorHandler) Errors() []error {
-	return h.errors
-}
-
-func (h *ErrorHandler) HasErrors() bool {
-	return len(h.errors) > 0
-}
-
-func (h *ErrorHandler) Error() string {
-	if !h.HasErrors() {
-		return ""
-	}
-	
-	var msgs []string
-	for _, err := range h.errors {
-		msgs = append(msgs, err.Error())
-	}
-	return strings.Join(msgs, "; ")
-}
-
-func (h *ErrorHandler) ToStrings() []string {
-	var strs []string
-	for _, err := range h.errors {
-		strs = append(strs, err.Error())
-	}
-	return strs
 }